Moves link tests
