+++ packages/link/__tests__/ios/link.ios.test.js
@@ -1,7 +1,7 @@
 import { Text } from "react-native";
 import test from "../shared";
-import Link, { TextLink } from "../../link";
+import Link, { TextLink } from "../../src/link";
 
-describe("Link test on android", () => {
+describe("Link test on ios", () => {
   test(Link, TextLink, Text);
 });
